% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myBoot.R
\name{myboot}
\alias{myboot}
\title{MATH4753COUR0004::myBoot()}
\usage{
myboot(iter = 10000, x, fun = "mean", alpha = 0.05, ...)
}
\arguments{
\item{iter}{The number of iterations you would like for the function to sample.}

\item{x}{A vector of values to pass in as the sample}

\item{fun}{The name of the function you would like to apply.}

\item{alpha}{The amount of confidence you would like, must be between 0 and 1}

\item{...}{Other arguments which will be passed to the hist() function for use with a plot.}
}
\value{
Returns the confidence interval, function, and original vector which was passed to x.
This will be returned in the form of a vector.
}
\description{
Facilitates a bootstrap procedure. This may be used to make a number of plots, point, and interval estimates
and useful output in the form of a list.
}
\examples{

set.seed(39)
sam=rnorm(25,mean=25,sd=10)
mb2_1=myboot2(iter=10000,x=sam,fun="mean",alpha=0.05,xlab="mean",col="red")
}
