% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myci.R
\name{myci}
\alias{myci}
\title{MATH4753COUR0004::myci()}
\usage{
myci(x = c())
}
\arguments{
\item{x}{A vector of values which you would like to create
a 95\% confidence interval  of the mean for.}
}
\description{
myci() creates a 95\% confidence interval of the mean given a vector x.
This is like a much simplified version of the t.test() function,
obviously with much less functionality.
}
\examples{
y=c(3,4,5,6,7)
confidence_interval=myci(y)
confidence_interval

}
