% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myncurve.R
\name{myncurve}
\alias{myncurve}
\title{MATH4753COUR0004::myncurve()}
\usage{
myncurve(mu = 0, sigma = 1, a)
}
\arguments{
\item{mu}{The mean of this normal distribution. Defaults to 0}

\item{sigma}{The standard deviation of this normal distribution. Defaults to 1}

\item{a}{The lower tail probability which the function begins at.}
}
\value{
This function does not return any values. Rather, it simply prints out the lower tail
probability of the given curve beginning at the value a
}
\description{
Creates a graph of a normal curve with mean mu and standard deviation sigma.
On this plot, it will also plot the lower tail area under the curve beginning at
a, and print this value out to the console.
}
\examples{
MATH4753COUR0004::myncurve(mu=10,sigma=5,a=6)
}
