% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myRandom.R
\name{myRandom}
\alias{myRandom}
\title{MATH4753COUR0004::myRandom()}
\usage{
myRandom(n, iter = 10, time = 0.5)
}
\arguments{
\item{n}{The size of the sample}

\item{iter}{How many times to choose the sample}

\item{time}{Time in between displaying each graph. May be useful with an animation}
}
\value{
- Generates new bar graphs every time seconds
}
\description{
MATH4753COUR0004::myRandom()
}
\examples{

myRandom(1000,30,1)
The above example generates a new bar graph every second for 30 seconds based on the data gathered from 1000 iterations
}
